refactor(models): refactor provider API request body preparation and message transformation
- Extracted image URL processing to a dedicated module
- Moved message transformation logic for Anthropic and Google providers to separate modules
- Added transformation for models without system role support
- Consolidated prepareRequestBody function with improved provider-specific handling
- Removed inline image processing and message transformation from provider-api.ts
- Improved error handling and logging for image processing
- Updated imports and exports to reflect new module structure

This refactor improves code modularity, readability, and maintainability by separating concerns and centralizing request body preparation logic.
