+++ packages/models/src/transform/no-system-role.ts
@@ -1,0 +1,11 @@
+export function transformMessagesForNoSystemRole(messages: any[]): any[] {
+	return messages.map((message) => {
+		if (message.role === "system") {
+			return {
+				...message,
+				role: "user",
+			};
+		}
+		return message;
+	});
+}
