+++ packages/models/src/transform/prepare-request-body.ts
@@ -1,0 +1,415 @@
+import {
+	models,
+	type ModelDefinition,
+	type ProviderModelMapping,
+} from "@/packages/models/src/models";
+
+import { transformAnthropicMessages } from "./anthropic-messages";
+import { transformGoogleMessages } from "./google-messages";
+import { transformMessagesForNoSystemRole } from "./no-system-role";
+
+import type { ProviderId } from "@/packages/models/src/providers";
+import type {
+	BaseMessage,
+	FunctionParameter,
+	OpenAIToolInput,
+	ToolChoiceType,
+	ProviderRequestBody,
+	OpenAIRequestBody,
+	OpenAIResponsesRequestBody,
+} from "@/packages/models/src/types";
+
+export async function prepareRequestBody(
+	usedProvider: ProviderId,
+	usedModel: string,
+	messages: BaseMessage[],
+	stream: boolean,
+	temperature: number | undefined,
+	max_tokens: number | undefined,
+	top_p: number | undefined,
+	frequency_penalty: number | undefined,
+	presence_penalty: number | undefined,
+	response_format: OpenAIRequestBody["response_format"],
+	tools?: OpenAIToolInput[],
+	tool_choice?: ToolChoiceType,
+	reasoning_effort?: "low" | "medium" | "high",
+	supportsReasoning?: boolean,
+	isProd = false,
+): Promise<ProviderRequestBody> {
+	// Check if the model supports system role
+	const modelDef = models.find((m) => m.id === usedModel);
+	const supportsSystemRole =
+		(modelDef as ModelDefinition)?.supportsSystemRole !== false;
+
+	// Transform messages if model doesn't support system role
+	let processedMessages = messages;
+	if (!supportsSystemRole) {
+		processedMessages = transformMessagesForNoSystemRole(messages);
+	}
+
+	// Start with a base structure that can be modified for each provider
+	const requestBody: any = {
+		model: usedModel,
+		messages: processedMessages,
+		stream: stream,
+	};
+	if (tools && tools.length > 0) {
+		requestBody.tools = tools;
+	}
+
+	if (tool_choice) {
+		requestBody.tool_choice = tool_choice;
+	}
+
+	switch (usedProvider) {
+		case "openai": {
+			// Override temperature to 1 for GPT-5 models (they only support temperature = 1)
+			let effectiveTemperature = temperature;
+			if (usedModel.startsWith("gpt-5")) {
+				effectiveTemperature = 1;
+			}
+
+			// Check if messages contain existing tool calls or tool results
+			// If so, use Chat Completions API instead of Responses API
+			const hasExistingToolCalls = messages.some(
+				(msg: any) => msg.tool_calls || msg.role === "tool",
+			);
+
+			// Check if the model supports responses API (default to true if reasoning is enabled)
+			const providerMapping = modelDef?.providers.find(
+				(p) => p.providerId === "openai",
+			);
+			const supportsResponsesApi =
+				process.env.USE_RESPONSES_API === "true" &&
+				(providerMapping as ProviderModelMapping)?.supportsResponsesApi !==
+					false;
+
+			if (supportsReasoning && supportsResponsesApi && !hasExistingToolCalls) {
+				// Transform to responses API format (only when no existing tool calls)
+				const responsesBody: OpenAIResponsesRequestBody = {
+					model: usedModel,
+					input: processedMessages,
+					reasoning: {
+						effort: reasoning_effort || "medium",
+						summary: "detailed",
+					},
+				};
+
+				// Add streaming support
+				if (stream) {
+					responsesBody.stream = true;
+				}
+
+				// Add tools support for responses API (transform format if needed)
+				if (tools && tools.length > 0) {
+					// Transform tools from chat completions format to responses API format
+					responsesBody.tools = tools.map((tool) => ({
+						type: "function" as const,
+						name: tool.function.name,
+						description: tool.function.description,
+						parameters: tool.function.parameters as FunctionParameter,
+					}));
+				}
+				if (tool_choice) {
+					responsesBody.tool_choice = tool_choice;
+				}
+
+				// Add optional parameters if they are provided
+				if (effectiveTemperature !== undefined) {
+					responsesBody.temperature = effectiveTemperature;
+				}
+				if (max_tokens !== undefined) {
+					responsesBody.max_output_tokens = max_tokens;
+				}
+
+				return responsesBody;
+			} else {
+				// Use regular chat completions format
+				if (stream) {
+					requestBody.stream_options = {
+						include_usage: true,
+					};
+				}
+				if (response_format) {
+					requestBody.response_format = response_format;
+				}
+
+				// Add optional parameters if they are provided
+				if (effectiveTemperature !== undefined) {
+					requestBody.temperature = effectiveTemperature;
+				}
+				if (max_tokens !== undefined) {
+					// GPT-5 models use max_completion_tokens instead of max_tokens
+					if (usedModel.startsWith("gpt-5")) {
+						requestBody.max_completion_tokens = max_tokens;
+					} else {
+						requestBody.max_tokens = max_tokens;
+					}
+				}
+				if (top_p !== undefined) {
+					requestBody.top_p = top_p;
+				}
+				if (frequency_penalty !== undefined) {
+					requestBody.frequency_penalty = frequency_penalty;
+				}
+				if (presence_penalty !== undefined) {
+					requestBody.presence_penalty = presence_penalty;
+				}
+				if (reasoning_effort !== undefined) {
+					requestBody.reasoning_effort = reasoning_effort;
+				}
+			}
+			break;
+		}
+		case "zai": {
+			if (stream) {
+				requestBody.stream_options = {
+					include_usage: true,
+				};
+			}
+			if (response_format) {
+				requestBody.response_format = response_format;
+			}
+
+			// Add optional parameters if they are provided
+			if (temperature !== undefined) {
+				requestBody.temperature = temperature;
+			}
+			if (max_tokens !== undefined) {
+				requestBody.max_tokens = max_tokens;
+			}
+			if (top_p !== undefined) {
+				requestBody.top_p = top_p;
+			}
+			if (frequency_penalty !== undefined) {
+				requestBody.frequency_penalty = frequency_penalty;
+			}
+			if (presence_penalty !== undefined) {
+				requestBody.presence_penalty = presence_penalty;
+			}
+			// ZAI/GLM models use 'thinking' parameter for reasoning instead of 'reasoning_effort'
+			if (supportsReasoning) {
+				requestBody.thinking = {
+					type: "enabled",
+				};
+			}
+			break;
+		}
+		case "xai":
+		case "groq":
+		case "deepseek":
+		case "perplexity":
+		case "novita":
+		case "moonshot":
+		case "alibaba":
+		case "nebius":
+		case "routeway":
+		case "custom": {
+			if (stream) {
+				requestBody.stream_options = {
+					include_usage: true,
+				};
+			}
+			if (response_format) {
+				requestBody.response_format = response_format;
+			}
+
+			// Add optional parameters if they are provided
+			if (temperature !== undefined) {
+				requestBody.temperature = temperature;
+			}
+			if (max_tokens !== undefined) {
+				requestBody.max_tokens = max_tokens;
+			}
+			if (top_p !== undefined) {
+				requestBody.top_p = top_p;
+			}
+			if (frequency_penalty !== undefined) {
+				requestBody.frequency_penalty = frequency_penalty;
+			}
+			if (presence_penalty !== undefined) {
+				requestBody.presence_penalty = presence_penalty;
+			}
+			if (reasoning_effort !== undefined) {
+				requestBody.reasoning_effort = reasoning_effort;
+			}
+			break;
+		}
+		case "anthropic": {
+			// Remove generic tool_choice that was added earlier
+			delete requestBody.tool_choice;
+
+			// Set max_tokens, ensuring it's higher than thinking budget when reasoning is enabled
+			const getThinkingBudget = (effort?: string) => {
+				if (!supportsReasoning) {
+					return 0;
+				}
+				if (!reasoning_effort) {
+					return 0;
+				}
+				switch (effort) {
+					case "low":
+						return 1024; // Anthropic minimum
+					case "high":
+						return 4000;
+					default:
+						return 2000; // medium or undefined
+				}
+			};
+			const thinkingBudget = getThinkingBudget(reasoning_effort);
+			const minMaxTokens = Math.max(1024, thinkingBudget + 1000);
+			requestBody.max_tokens = max_tokens ?? minMaxTokens;
+			requestBody.messages = await transformAnthropicMessages(
+				processedMessages.map((m) => ({
+					...m, // Preserve original properties for transformation
+					role:
+						m.role === "assistant"
+							? "assistant"
+							: m.role === "system"
+								? "user"
+								: m.role === "tool"
+									? "user" // Tool results become user messages in Anthropic
+									: "user",
+					content: m.content,
+					tool_calls: m.tool_calls, // Include tool_calls for transformation
+				})),
+				isProd,
+			);
+
+			// Transform tools from OpenAI format to Anthropic format
+			if (tools && tools.length > 0) {
+				requestBody.tools = tools.map((tool) => ({
+					name: tool.function.name,
+					description: tool.function.description,
+					input_schema: tool.function.parameters,
+				}));
+			}
+
+			// Handle tool_choice parameter - transform OpenAI format to Anthropic format
+			if (tool_choice) {
+				if (
+					typeof tool_choice === "object" &&
+					tool_choice.type === "function"
+				) {
+					// Transform OpenAI format to Anthropic format
+					requestBody.tool_choice = {
+						type: "tool",
+						name: tool_choice.function.name,
+					};
+				} else if (tool_choice === "auto") {
+					// "auto" is the default behavior for Anthropic, omit it
+					// Anthropic doesn't need explicit "auto" tool_choice
+				} else if (tool_choice === "none") {
+					// "none" should work as-is
+					requestBody.tool_choice = tool_choice;
+				} else {
+					// Other string values (though not standard)
+					requestBody.tool_choice = tool_choice;
+				}
+			}
+
+			// Enable thinking for reasoning-capable Anthropic models when reasoning_effort is specified
+			if (supportsReasoning && reasoning_effort) {
+				requestBody.thinking = {
+					type: "enabled",
+					budget_tokens: thinkingBudget,
+				};
+			}
+
+			// Add optional parameters if they are provided
+			if (temperature !== undefined) {
+				requestBody.temperature = temperature;
+			}
+			if (top_p !== undefined) {
+				requestBody.top_p = top_p;
+			}
+			if (frequency_penalty !== undefined) {
+				requestBody.frequency_penalty = frequency_penalty;
+			}
+			if (presence_penalty !== undefined) {
+				requestBody.presence_penalty = presence_penalty;
+			}
+			break;
+		}
+		case "google-ai-studio": {
+			delete requestBody.model; // Not used in body
+			delete requestBody.stream; // Stream is handled via URL parameter
+			delete requestBody.messages; // Not used in body for Google providers
+			delete requestBody.tool_choice; // Google doesn't support tool_choice parameter
+
+			requestBody.contents = await transformGoogleMessages(
+				processedMessages,
+				isProd,
+			);
+
+			// Transform tools from OpenAI format to Google format
+			if (tools && tools.length > 0) {
+				requestBody.tools = [
+					{
+						functionDeclarations: tools.map((tool: any) => {
+							// Remove additionalProperties and $schema from parameters as Google doesn't accept them
+							const {
+								additionalProperties: _additionalProperties,
+								$schema: _$schema,
+								...cleanParameters
+							} = tool.function.parameters || {};
+							return {
+								name: tool.function.name,
+								description: tool.function.description,
+								parameters: cleanParameters,
+							};
+						}),
+					},
+				];
+			}
+
+			requestBody.generationConfig = {};
+
+			// Add optional parameters if they are provided
+			if (temperature !== undefined) {
+				requestBody.generationConfig.temperature = temperature;
+			}
+			if (max_tokens !== undefined) {
+				requestBody.generationConfig.maxOutputTokens = max_tokens;
+			}
+			if (top_p !== undefined) {
+				requestBody.generationConfig.topP = top_p;
+			}
+
+			// Enable thinking/reasoning content exposure for Google models that support reasoning
+			if (supportsReasoning) {
+				requestBody.generationConfig.thinkingConfig = {
+					includeThoughts: true,
+				};
+			}
+
+			break;
+		}
+		case "inference.net":
+		case "together.ai": {
+			if (usedModel.startsWith(`${usedProvider}/`)) {
+				requestBody.model = usedModel.substring(usedProvider.length + 1);
+			}
+
+			// Add optional parameters if they are provided
+			if (temperature !== undefined) {
+				requestBody.temperature = temperature;
+			}
+			if (max_tokens !== undefined) {
+				requestBody.max_tokens = max_tokens;
+			}
+			if (top_p !== undefined) {
+				requestBody.top_p = top_p;
+			}
+			if (frequency_penalty !== undefined) {
+				requestBody.frequency_penalty = frequency_penalty;
+			}
+			if (presence_penalty !== undefined) {
+				requestBody.presence_penalty = presence_penalty;
+			}
+			break;
+		}
+	}
+
+	return requestBody;
+}
